getting asknews working within limits
appears that there are some pretty strict limits on what asknews is providing for free to metaculus. The template bot doesn't conform with these, so needed to make changes. ALSO needed to change the model in the predictiontools asknews. This function might need to be moved into my own script.